feat(dsg) final work before merge
