+++ helm/charts/amplication/values-amit.yaml
@@ -1,13 +1,8 @@
 global:
   pvc:
-    name: build-artifacts
-    storageRequest: 5Gi
-    storageName: efs-sc
+    name: amit-efs-claim
   pv:
-    name: build-artifacts-amit
-    storageCapacity: 5Gi
-    handle: fs-03a0a5ba30d788595::fsap-08494d4659781878b
-
+    name: efs-pv-amit
   nodeSelector:
     role: custom-envs
   tolerations:
@@ -33,10 +28,6 @@ amplication-client:
       NX_REACT_APP_STORAGE_GATEWAY_URL: https://server-amit.amplication-dev.com
 
 amplication-server:
-  autoscaling:
-    minReplicas: 4
-    maxReplicas: 99
-  replicaCount: 4
   ingress:
     hostname: server-amit.amplication-dev.com
   config:
@@ -45,19 +36,13 @@ amplication-server:
       CLIENT_HOST: "https://app-amit.amplication-dev.com"
 
       #KAFKA
+      GENERATE_PULL_REQUEST_TOPIC: git.internal.pull-request.request.0
       KAFKA_BROKERS: 'amplication-amit-cp-kafka-headless:9092'
       KAFKA_CLIENT_ID: amplication-server
       KAFKA_GROUP_ID: amplication-server-group
 
-      CREATE_PR_REQUEST_TOPIC: git.internal.create-pr.request.1
-      CREATE_PR_SUCCESS_TOPIC: git.internal.create-pr.success.1
-      CREATE_PR_FAILURE_TOPIC: git.internal.create-pr.failure.1
-
-      CODE_GENERATION_REQUEST_TOPIC: build.internal.code-generation.request.1
-      CODE_GENERATION_SUCCESS_TOPIC: build.internal.code-generation.success.1
-      CODE_GENERATION_FAILURE_TOPIC: build.internal.code-generation.failure.1
-
-      DSG_LOG_TOPIC: build.internal.dsg-log.1
+      #EFS
+      BASE_BUILDS_FOLDER: /amplication-data/build-data
 
       #GCP - APPS Builder
 
@@ -76,8 +61,6 @@ amplication-server:
       GCP_APPS_DOMAIN: staging.amplication.app
 
       DOCKER_HOST: tcp://docker:2375
-
-      DSG_RUNNER_URL: http://webhook-eventsource-svc.amit-dsg.svc.cluster.local:12000/example
 
       #GitHub App
       GITHUB_APP_APP_ID: "215215"
@@ -124,14 +107,8 @@ amplication-git-pull-request-service:
       KAFKA_BROKERS: 'amplication-amit-cp-kafka-headless:9092'
       KAFKA_CLIENT_ID: git-pull-request-service
       KAFKA_GROUP_ID: git-pull-request-service-group
-
-      CREATE_PR_REQUEST_TOPIC: git.internal.create-pr.request.1
-      CREATE_PR_SUCCESS_TOPIC: git.internal.create-pr.success.1
-      CREATE_PR_FAILURE_TOPIC: git.internal.create-pr.failure.1
-
-      BUILD_ARTIFACTS_BASE_FOLDER: /amplication-data/build-data
-      BUILD_ARTIFACTS_CODE_FOLDER: code
-
+      GENERATE_PULL_REQUEST_TOPIC: git.internal.pull-request.request.0
+      BASE_BUILDS_FOLDER: /amplication-data/build-data
       GITHUB_APP_APP_ID: "215215"
 
 amplication-storage-gateway:
@@ -141,5 +118,5 @@ amplication-storage-gateway:
     env:
       KAFKA_BROKERS: 'amplication-amit-cp-kafka-headless:9092'
       BASE_BUILDS_FOLDER: /amplication-data/build-data
-      BUILD_OUTPUT_FOLDER: code
       GITHUB_APP_APP_ID: "215215"
+
