+++ packages/amplication-data-service-generator/src/server/create-server.ts
@@ -24,6 +24,8 @@ import { createDockerComposeDBFile } from "./docker-compose/create-docker-compos
 import { createDockerComposeFile } from "./docker-compose/create-docker-compose";
 import pluginWrapper from "../plugin-wrapper";
 import { createLog } from "../create-log";
+import { createUserInfo } from "./auth/user-info/create-user-info";
+import { createTokenPayload } from "./auth/token/create-token-payload-interface";
 
 const STATIC_DIRECTORY = path.resolve(__dirname, "static");
 
@@ -71,6 +73,17 @@ async function createServerInternal(
   const dtoModules = createDTOModules(dtos);
   const resourcesModules = await createResourcesModules(entities, logger);
 
+  await createLog({ level: "info", message: "Creating User Info..." });
+  logger.info("Creating User Info...");
+  const userInfo = await createUserInfo();
+
+  await createLog({
+    level: "info",
+    message: "Creating Token Payload Interface...",
+  });
+  logger.info("Token Payload Interface...");
+  const tokenPayloadInterface = await createTokenPayload();
+
   await createLog({ level: "info", message: "Creating Auth module..." });
   logger.info("Creating Auth module...");
   const authModules = await createAuthModules();
@@ -102,6 +115,8 @@ async function createServerInternal(
     ...swagger,
     ...appModule,
     ...seedModule,
+    userInfo,
+    tokenPayloadInterface,
     ...authModules,
     ...messageBrokerModules,
   ];
